Added high score functionality but without storing them post-exit

+++ source/core/src/main/com/deco2800/game/screens/MainGameScreen.java
@@ -44,7 +44,7 @@ public class MainGameScreen extends ScreenAdapter {
   private final PhysicsEngine physicsEngine;
 
   private final long timeStarted = System.currentTimeMillis();
-  public static long timeScore = 0;
+  public static int timeScore = 0;
 
   public MainGameScreen(GdxGame game) {
     this.game = game;
@@ -89,7 +89,7 @@ public class MainGameScreen extends ScreenAdapter {
       logger.info("Show Death Screen");
       game.setScreen(GdxGame.ScreenType.DEATH_SCREEN);
     }
-    timeScore = (System.currentTimeMillis() - timeStarted) / 1000;
+    timeScore = (int) ((System.currentTimeMillis() - timeStarted) / 1000);
   }
 
   @Override
